refactor: relocate tag-gap provenance out of deploy so extension build stops depending on deploy
detect_tag_gap / warn_tag_gap / TagGap warn when a component's HEAD is ahead of
its latest release tag. Their own doc says they're 'used by both build and deploy
commands' — a shared build-provenance advisory that only happened to live under
deploy::provenance, forcing the extension build path to import crate::deploy.

Relocate the tag-gap trio (+ private format_tag_gap helper) to a core-level
crate::tag_gap module (deps only component/engine::command/release). deploy's
provenance re-exports detect_tag_gap/warn_tag_gap so its internal consumers
(orchestration::preflight) are unchanged; extension/build/mod.rs now imports
crate::tag_gap directly.

This severs one of the extension build path's two deploy edges (the provenance
one). The remaining deploy edge — deploy::permissions::fix_local_permissions — is
a separate concern for a later cut. Continues draining extension's upward edges
toward a future homeboy-extension extraction.

Refs #8425
